requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
gunicorn
pkg-resources
pytz
sqlparse
typing-extensions
whitenoise
 
